refactor: extract payment helpers to utils module
