+++ renderer/reducers/payment/utils.js
@@ -1,5 +1,10 @@
-import { decodePayReq, getNodeAlias } from '@zap/utils/crypto'
+import config from 'config'
+import { decodePayReq, getNodeAlias, getTag, generatePreimage } from '@zap/utils/crypto'
 import { convert } from '@zap/utils/btc'
+import { sha256digest } from '@zap/utils/sha256'
+import { DEFAULT_CLTV_DELTA, KEYSEND_PREIMAGE_TYPE } from './constants'
+
+const PAYMENT_TIMEOUT = config.invoices.paymentTimeout
 
 /**
  * decoratePayment - Decorate payment object with custom/computed properties.
@@ -56,5 +61,62 @@ export const decoratePayment = (payment, nodes = []) => {
   return {
     ...payment,
     ...decoration,
+  }
+}
+
+/**
+ * getPaymentConfig - Get common payment settings.
+ *
+ * @returns {object} common payment settings
+ */
+export const getPaymentConfig = () => {
+  return {
+    allowSelfPayment: true,
+    timeoutSeconds: PAYMENT_TIMEOUT,
+  }
+}
+
+/**
+ * prepareKeysendPayload - Prepare a keysend payment.
+ *
+ * @param  {string} pubkey   Pubkey
+ * @param  {number} amt      Amount in satoshi
+ * @param  {number} feeLimit Fee limit (sats)
+ * @returns {object} Keysnd payment
+ */
+export const prepareKeysendPayload = (pubkey, amt, feeLimit) => {
+  const preimage = generatePreimage()
+
+  return {
+    ...getPaymentConfig(),
+    dest: Buffer.from(pubkey, 'hex'),
+    feeLimit: feeLimit ? { fixed: feeLimit } : null,
+    paymentHash: sha256digest(preimage),
+    amt,
+    finalCltvDelta: DEFAULT_CLTV_DELTA,
+    destCustomRecords: {
+      [KEYSEND_PREIMAGE_TYPE]: preimage,
+    },
+  }
+}
+
+/**
+ * prepareKeysendPayload - Prepare a bolt11 payment.
+ *
+ * @param  {string} payReq   Payment request
+ * @param  {number} amt      Amount in satoshi
+ * @param  {number} feeLimit Fee limit (sats)
+ * @returns {object} bolt11 payment
+ */
+export const prepareBolt11Payload = (payReq, amt, feeLimit) => {
+  const invoice = decodePayReq(payReq)
+  const { millisatoshis } = invoice
+
+  return {
+    ...getPaymentConfig(),
+    paymentRequest: invoice.paymentRequest,
+    paymentHash: getTag(invoice, 'payment_hash'), // hash is not needed in the payload but store for convienience.
+    feeLimit: feeLimit ? { fixed: feeLimit } : null,
+    amt: millisatoshis ? null : amt,
   }
 }
